temporary modification of makedir

+++ jetset/output.py
@@ -21,20 +21,21 @@ def makedir(out_dir,clean_work_dir=True):
     """
     creates a directory
     """
-    if os.path.isdir(out_dir):
-        Warning ("directory %s already existing"%(out_dir))
-        if clean_work_dir==True:
-            Warning ('removing existing dir')
-            shutil.rmtree(out_dir)
-            os.mkdir(out_dir)
-            Warning ('the directory %s has been created' % (out_dir))
-    else:
-        if os.path.isfile(out_dir):
-            Warning ("a file with the same name of dir=%s, exists"%out_dir)
-            Warning ("select a differn name")
+    if not os.path.exists(out_dir):
+        if os.path.isdir(out_dir):
+            Warning ("directory %s already existing"%(out_dir))
+            if clean_work_dir==True:
+                Warning ('removing existing dir')
+                shutil.rmtree(out_dir)
+                os.mkdir(out_dir)
+                Warning ('the directory %s has been created' % (out_dir))
         else:
-            os.mkdir(out_dir)
-            Warning ('the directory %s has been created'%(out_dir))
+            if os.path.isfile(out_dir):
+                Warning ("a file with the same name of dir=%s, exists"%out_dir)
+                Warning ("select a differn name")
+            else:
+                os.mkdir(out_dir)
+                Warning ('the directory %s has been created'%(out_dir))
 
 
 class WorkPlace(object):
